api/requirements.txt
flask
flask-cors
python-dotenv
python-jose
Flask-PyMongo
pymongo